refactor(primarynav): move to recipe
