+++ .storybook/recipes/GlobalHeader/GlobalHeader.tsx
@@ -2,14 +2,12 @@ import clsx from 'clsx';
 import type { ReactNode } from 'react';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import styles from './GlobalHeader.module.css';
-import { Button, Icon, Heading } from '../../../src';
-import Link from '../../../src/components/Link';
-import Popover from '../../../src/components/Popover';
-import PrimaryNav from '../../../src/components/PrimaryNav';
+import { Button, Icon, Heading, Link, Popover } from '../../../src';
 
 import breakpoint from '../../../src/design-tokens/tier-1-definitions/breakpoints';
 import { EdsThemeColorIconNeutralDefaultInverse } from '../../../src/tokens-dist/ts/colors';
 import { NotificationLists } from '../NotificationListPopover/NotificationListPopover';
+import { PrimaryNav } from '../PrimaryNav/PrimaryNav';
 
 type HeaderProps = {
   /**
