+++ .storybook/recipes/PrimaryNav/PrimaryNav.module.css
@@ -1,4 +1,4 @@
-@import '../../design-tokens/mixins.css';
+@import '../../../src/design-tokens/mixins.css';
 
 /*------------------------------------*\
     # PRIMARY NAVIGATION ITEM
