--- src/components/PrimaryNav/index.ts
@@ -1,1 +1,0 @@
-export { PrimaryNav as default } from './PrimaryNav';
