+++ src/index.ts
@@ -100,15 +100,3 @@ export { default as Tooltip } from './components/Tooltip';
  */
 export type { IconName } from './components/Icon';
 export type { BannerProps } from './components/Banner';
-
-/**
- * Project Overview Pilot components.
- *
- * These are unused at the moment -- we can re-add
- * the exports once we confirm this package is set
- * up correctly to be tree shaken.
- * In the meantime, leaving them commented out
- * ensures we don't unintentionally blow up the bundle size of
- * consuming apps.
- */
-// export { default as PrimaryNav } from './components/PrimaryNav';
